+++ VoiceCraft.Core/Interfaces/IVisible.cs
@@ -4,6 +4,6 @@ namespace VoiceCraft.Core.Interfaces
 {
     public interface IVisible
     {
-        bool Visibility(VoiceCraftEntity from, VoiceCraftEntity to, uint effectBitmask);
+        bool Visibility(VoiceCraftEntity from, VoiceCraftEntity to, ushort effectBitmask);
     }
 }
